+++ changelog.md
@@ -1,2 +1,2 @@
 ### Landmines 1.0.7
-- Initial release of Landmines for NeoForge
+- Initial release of Landmines for Minecraft 1.20.5 & 1.20.6